requirements.txt
requests==2.32.5
aiohttp==3.13.3
httpx==0.28.1

openai==2.20.0
anthropic>=0.79.0
google-genai==1.62.0
protobuf==6.33.5

fastmcp==3.0.2
redis[hiredis]==7.1.1
fastapi==0.128.7

langchain-community==0.4.1
langchain-openai==1.1.10
langchain-chroma==1.1.0
faiss-cpu==1.13.2
psycopg2-binary>=2.9.11
presidio-analyzer==2.2.360
presidio_anonymizer==2.2.360
pydantic==2.12.5
python-jose==3.5.0